order-main/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
psycopg2==2.9.5
python-dotenv==1.0.0
uWSGI==2.0.21
drf-yasg==1.21.5
requests==2.28.2